+++ src/index.ts
@@ -26,14 +26,10 @@ export {
     StringFunction,
 } from './types';
 export {
-    SessionPool,
-    Session,
+    TableSession,
     CreateTableSettings,
     AlterTableSettings,
     DropTableSettings,
-    BeginTransactionSettings,
-    CommitTransactionSettings,
-    RollbackTransactionSettings,
     DescribeTableSettings,
     PrepareQuerySettings,
     ExecuteQuerySettings,
@@ -54,9 +50,8 @@ export {
     CompactionPolicy,
     ExecutionPolicy,
     CachingPolicy,
-    OperationParams,
-    AUTO_TX,
-} from './table';
+
+} from './table/table-client';
 export {
     MakeDirectorySettings,
     RemoveDirectorySettings,
@@ -78,3 +73,9 @@ export {
 export {ISslCredentials} from './ssl-credentials';
 export {withRetries, RetryParameters} from './retries';
 export {YdbError, StatusCode} from './errors';
+export {SessionPool} from "./table/session-pool";
+export {RollbackTransactionSettings} from "./table/session";
+export {CommitTransactionSettings} from "./table/session";
+export {BeginTransactionSettings} from "./table/session";
+export {OperationParams} from "./table/session";
+export {AUTO_TX} from "./table/session";