+++ src/table/index.ts
@@ -1,0 +1,3 @@
+export * from './session-pool';
+export * from './session';
+export * from './table-client';